all setup of register and bugs fixis

+++ src/utils/cloudinary.js
@@ -4,23 +4,26 @@ import fs from "fs";
     // Configuration
     cloudinary.config({ 
         cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY , 
+        api_key: process.env.CLOUDINARY_API_KEY, 
         api_secret: process.env.CLOUDINARY_API_SECRET
     });
 
 
-const uploadOnClodinary = async(localFilePath)=>{
+const uploadOnCloudinary = async(localFilePath)=>{
     try {
+
         if (!localFilePath) return null;
         //upload the file on cloudnary
         const response = await cloudinary.uploader.upload(localFilePath,{
             resource_type:"auto"
         })
+        fs.unlinkSync(localFilePath)
         return response;
     } catch (error) {
-       fs.unlinkSync(localFilePath)
-       return null;
-    }
+   console.log("Cloudinary Upload Error:", error);
+   if (localFilePath) fs.unlinkSync(localFilePath);
+   return null;
+}
 } 
 
-export {uploadOnClodinary}
+export {uploadOnCloudinary}
